-able to pass sample player action input from front-end to TableService.java
-currently debugging tableService methods.

+++ junopoker/src/main/java/com/v1/junopoker/model/Table.java
@@ -24,7 +24,7 @@ public class Table {
     private int bigBlind;
     private int smallBlind;
     private int dealerButton;
-    private int currentBet;
+    private float currentBet;
     private boolean gameIsRunning;
     private DeckService deckService;
